Feat: 컴포넌트 분리 & 로딩 UI
