+++ src/components/Loading.tsx
@@ -1,0 +1,19 @@
+import { ReactComponent as SpinnerIcon } from "../assets/icon/spinner.svg";
+
+const Loading = () => {
+  return (
+    <div className="container">
+      <div className="loading-wrapper">
+        <SpinnerIcon
+          width="80px"
+          heigth="80px"
+          fill="#fff"
+          className="animate"
+        />
+        <p>Loading</p>
+      </div>
+    </div>
+  );
+};
+
+export default Loading;
